# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Sistema de Frequencia Avance",
    "version": "1.1.0",
    "description": "Documentacao oficial da API versionada em /api/v1 com envelope padrao de resposta."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versao atual"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": ["success", "data", "message", "errors"],
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "nullable": true,
            "description": "Payload principal de sucesso."
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string", "nullable": true },
                "location": { "type": "string", "nullable": true },
                "message": { "type": "string" },
                "code": { "type": "string", "nullable": true },
                "value": { "nullable": true }
              }
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["usuario", "senha"],
        "properties": {
          "usuario": { "type": "string", "example": "admin" },
          "senha": { "type": "string", "example": "Senha@12345" }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        }
      },
      "UsuarioRegisterRequest": {
        "type": "object",
        "required": ["nome", "usuario", "senha"],
        "properties": {
          "nome": { "type": "string", "example": "Joao Silva" },
          "usuario": { "type": "string", "example": "joao.silva" },
          "senha": { "type": "string", "example": "Senha@12345" }
        }
      },
...
      "CalendarioEventoRequest": {
        "type": "object",
        "required": ["titulo", "data", "tipo"],
        "properties": {
          "titulo": { "type": "string", "example": "Reuniao com responsaveis" },
          "data": { "type": "string", "format": "date", "example": "2026-05-10" },
          "tipo": {
            "type": "string",
            "enum": ["feriado", "reuniao", "prova", "evento", "conselho", "recesso", "aula-especial", "observacao"]
          },
          "descricao": { "type": "string" },
          "cor": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$" }
        }
      }
    },
    "responses": {
      "Api200": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "ok": {
                "value": {
                  "success": true,
                  "data": { "id": 1, "nome": "Joao" },
                  "message": "Operacao realizada com sucesso.",
                  "errors": null
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erro de validacao",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "validation": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Erro de validacao nos dados enviados.",
                  "errors": [
                    {
                      "field": "page",
                      "location": "query",
                      "message": "page deve ser um inteiro maior que 0.",
                      "value": "0"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Nao autenticado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "missing-token": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Token de acesso ausente.",
                  "errors": [
                    { "code": "AUTH_TOKEN_MISSING" }
                  ]
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "Limite de requisicoes excedido",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiResponse" },
            "examples": {
              "too-many-requests": {
                "value": {
                  "success": false,
                  "data": null,
                  "message": "Muitas tentativas de autenticacao. Tente novamente mais tarde.",
                  "errors": [
                    {
                      "code": "RATE_LIMIT_EXCEEDED",
                      "retryAfterSeconds": 900,
                      "path": "/api/v1/auth/login"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Cadastra usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioRegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Api200" },
          "409": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Realiza login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Solicita recuperacao de senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["usuario"],
                "properties": {
                  "usuario": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Renova sessao com refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Encerra sessao",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/calendario": {
      "get": {
        "tags": ["Calendario"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Lista eventos do calendario",
        "parameters": [
          { "name": "mes", "in": "query", "schema": { "type": "string", "example": "2026-05" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "post": {
        "tags": ["Calendario"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Cria evento do calendario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CalendarioEventoRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/calendario/{eventoId}": {
      "get": {
        "tags": ["Calendario"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Busca evento por ID",
        "parameters": [
          { "name": "eventoId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "put": {
        "tags": ["Calendario"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Atualiza evento",
        "parameters": [
          { "name": "eventoId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CalendarioEventoRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Api200" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      },
      "delete": {
        "tags": ["Calendario"],
        "security": [{ "bearerAuth": [] }],
        "summary": "Remove evento",
        "parameters": [
          { "name": "eventoId", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "204": { "description": "Sem conteudo" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/Api200" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    }
  }
}